+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
+- Add Theme Provider + Theme Switcher on Header for development (should be in settings in future)
+
 ### Changed
 
 ### Removed